Move timing into io/ module
