+++ src/core/mod.rs
@@ -1,4 +1,3 @@
 pub mod model;
 pub mod physics;
 pub mod scheduler;
-pub mod timing;
