+++ src/io/mod.rs
@@ -1,2 +1,3 @@
 pub mod railml_rollingstock;
+pub mod timing;
 pub mod xml_util;
